pyproject.toml
[tool.poetry]
name = "lxdbackup"
version = "0.1.0"
description = ""
authors = ["Jon Clayton <m87tech@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.7"
pydantic = "^1.8.2"
paramiko = "^2.7.2"
PyYAML = "^5.4.1"
pylxd = "^2.3.0"
confuse = "^1.4.0"

[tool.poetry.dev-dependencies]
pytest = "^5.2"
black = "^21.6b0"
pytest-cov = "^2.12.1"
flake8 = "^3.9.2"
eradicate = "^2.0.0"
mypy = "^0.902"
nox = "^2021.6.12"
pylint = "^2.8.3"
pydocstyle = "^6.1.1"
pycodestyle = "^2.7.0"
rope = "^0.19.0"
pytest-cases = "^3.6.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
